+++ crates/jstz_oracle_node/Cargo.toml
@@ -14,6 +14,7 @@ description.workspace = true
 anyhow.workspace = true
 futures.workspace = true
 futures-core.workspace = true
+pin-project.workspace = true
 tokio.workspace = true
 regex.workspace = true
 serde.workspace = true
