+++ crates/jstz_oracle_node/src/lib.rs
@@ -6,7 +6,6 @@ use jstz_utils::KeyPair;
 use octez::r#async::endpoint::Endpoint;
 use serde::Serialize;
 mod data_provider;
-mod filtered_log_stream;
 pub mod node;
 pub mod relay;
 pub mod request;
