+++ crates/jstz_utils/Cargo.toml
@@ -12,8 +12,17 @@ description = "Utility functions"
 
 [dependencies]
 anyhow.workspace = true
+futures.workspace = true
+futures-core.workspace = true
+jstz_core   = { path = "../jstz_core" }
 jstz_crypto = { path = "../jstz_crypto" }
 reqwest.workspace = true
+regex.workspace = true
 serde.workspace = true
 serde_json.workspace = true
+tempfile.workspace = true
 tokio.workspace = true
+tokio-util.workspace = true
+
+[dev-dependencies]
+futures-util.workspace = true
